fixed shape bug

+++ src/pokemon_classification/data.py
@@ -39,7 +39,7 @@ class PokemonDataset(Dataset):
 
     def __getitem__(self, idx):
         img_path = os.path.join(self.img_dir, self.img_labels.iloc[idx, 0])
-        image = decode_image(img_path, mode="RGB").unsqueeze(0)
+        image = decode_image(img_path, mode="RGB").unsqueeze(0).type(torch.float32)
         label = self.img_labels.iloc[idx, 1]
         if self.transform:
             image = self.transform(image)
